fixed landing page, added login,signup page
