+++ src/components/signupPage.jsx
@@ -1,0 +1,121 @@
+import React from 'react';
+import Avatar from '@material-ui/core/Avatar';
+import Button from '@material-ui/core/Button';
+import CssBaseline from '@material-ui/core/CssBaseline';
+import TextField from '@material-ui/core/TextField';
+import Link from '@material-ui/core/Link';
+import Grid from '@material-ui/core/Grid';
+import Box from '@material-ui/core/Box';
+import LockOutlinedIcon from '@material-ui/icons/LockOutlined';
+import Typography from '@material-ui/core/Typography';
+import { makeStyles } from '@material-ui/core/styles';
+import Container from '@material-ui/core/Container';
+import logo from '../static/library.png';
+
+const useStyles = makeStyles((theme) => ({
+  paper: {
+    marginTop: theme.spacing(8),
+    display: 'flex',
+    flexDirection: 'column',
+    alignItems: 'center',
+  },
+  avatar: {
+    margin: theme.spacing(1),
+    fontSize: '30px',
+    fontWeight: '600',
+    color: '#23466c',
+    display: 'flex',
+    justifyContent: 'center',
+    alignItems: 'center',
+  },
+  form: {
+    width: '100%',
+    marginTop: theme.spacing(3),
+  },
+  submit: {
+    margin: theme.spacing(3, 0, 2),
+  },
+}));
+
+export default function SignUp(props) {
+  const classes = useStyles();
+
+  return (
+    <Container component='main' maxWidth='xs'>
+      <CssBaseline />
+      <div className={classes.paper}>
+        <Avatar className={classes.avatar}>
+          <LockOutlinedIcon />
+        </Avatar>
+        <Typography className={classes.avatar} style={{ cursor: 'pointer' }}>
+          Bookmate <img src={logo} alt='' style={{ height: '30px' }} />
+        </Typography>
+        <Grid container spacing={2} style={{ margin: '10px' }}>
+          <Grid item xs={12} sm={6}>
+            <TextField
+              autoComplete='fname'
+              name='firstName'
+              variant='outlined'
+              required
+              fullWidth
+              id='firstName'
+              label='First Name'
+              autoFocus
+            />
+          </Grid>
+          <Grid item xs={12} sm={6}>
+            <TextField
+              variant='outlined'
+              required
+              fullWidth
+              id='lastName'
+              label='Last Name'
+              name='lastName'
+              autoComplete='lname'
+            />
+          </Grid>
+          <Grid item xs={12}>
+            <TextField
+              variant='outlined'
+              required
+              fullWidth
+              id='email'
+              label='Email Address'
+              name='email'
+              autoComplete='email'
+            />
+          </Grid>
+          <Grid item xs={12}>
+            <TextField
+              variant='outlined'
+              required
+              fullWidth
+              name='password'
+              label='Password'
+              type='password'
+              id='password'
+              autoComplete='current-password'
+            />
+          </Grid>
+          {/* <Grid item xs={12}>
+              <FormControlLabel
+                control={<Checkbox value='allowExtraEmails' color='primary' />}
+                label='I want to receive inspiration, marketing promotions and updates via email.'
+              />
+            </Grid> */}
+        </Grid>
+        <Button type='submit' fullWidth variant='contained' color='primary' className={classes.submit}>
+          Sign Up
+        </Button>
+        <Grid container justify='flex-end'>
+          <Grid item>
+            <Link href='#' variant='body2' onClick={() => props.onSignInClick()}>
+              Already have an account? Sign in
+            </Link>
+          </Grid>
+        </Grid>
+      </div>
+      <Box mt={5}>{/* <Copyright /> */}</Box>
+    </Container>
+  );
+}
